% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvd.R
\name{hvd}
\alias{hvd}
\title{Hilbert Vibration Decomposition}
\usage{
hvd(x, n, fp)
}
\arguments{
\item{x}{A numeric vector representing the input signal.}

\item{n}{The number of components to decompose the signal into.}

\item{fp}{The cutoff frequency for the lowpass filter.}
}
\value{
A list containing the following components:
\item{Y}{A matrix of the decomposed signal components.}
\item{A}{A matrix of the amplitudes of the decomposed signal components.}
\item{om_r}{A matrix of the angular frequencies of the decomposed signal components.}
\item{dev}{A vector of the relative standard deviation of the components.}
}
\description{
This function performs a Hilbert Vibration Decomposition (HVD) on a signal.
It decomposes the input signal into \code{n} components and returns their amplitudes,
frequencies, and phase shifts.
}
\examples{
om <- 0.2 + 0.12 * cos(0.4 * (0:1023))
x <- cos(cumsum(om))
result <- hvd(x, 3, 0.02)

Y <- result$Y
A <- result$A
F_r <- result$om_r
dev <- result$dev

par(mfrow = c(2, 1))
plot(x, type = 'l', xlim = c(400, 600), ylim = c(-1.1, 1.1), ylab = 'Initial signal')
plot(Y[,1], type = 'l', col=1, xlim = c(400, 600), ylim = c(-1.1, 1.1), xlab = 'Points', ylab = 'Signal Components')
lines(Y[,2], col = 2)
lines(Y[,3], col = 3)

library(stats)
plot(spec.pgram(x),log="dB")
}
