% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl}
\alias{pl}
\title{Plot Decomposed Components and Hilbert Spectrum}
\usage{
pl(Y, A, F_r, Fs)
}
\arguments{
\item{Y}{A matrix of decomposed components obtained from HVD.}

\item{A}{A matrix of component envelopes.}

\item{F_r}{A matrix of component relative angular frequencies.}

\item{Fs}{The sampling frequency.}
}
\value{
This function does not return a value but produces plots.
}
\description{
This function plots the decomposed components, their instantaneous frequencies,
envelopes, and the sum of components using the Hilbert spectrum.
}
\examples{
om <- 0.2 + 0.12 * cos(0.4 * (0:1023))
x <- cos(cumsum(om))
result <- hvd(x, 3, 0.02)
pl(result$Y, result$A, result$om_r, 2 * pi)

}
\references{
Feldman, M. (2011). HILBERT TRANSFORM APPLICATION IN MECHANICAL VIBRATION. John Wiley & Sons.
}
